Initialize Evaluator class
