+++ presidio_evaluator/evaluation/__init__.py
@@ -1,5 +1,6 @@
-from .model_error import ModelError
+from .evaluator_objects import SpanOutput, TokenOutput, ModelPrediction
+from .sample_error import SampleError
 from .evaluation_result import EvaluationResult
 from .evaluator import Evaluator
 
-__all__ = ["ModelError", "EvaluationResult", "Evaluator"]
+__all__ = ["SpanOutput", "TokenOutput", "ModelPrediction", "SampleError", "EvaluationResult", "Evaluator"]
